# Disco-Dancer/payment-vault/api/docs/swagger.yaml
definitions:
  app.ErrorResponse:
    properties:
      message:
        example: some error happened
        type: string
      type:
        example: error
        type: string
    type: object
  app.HealthCheckResponse:
    properties:
      configfilename:
        example: appsettings.dv.json
        type: string
      data: {}
      databasestatus:
        example: Up
        type: string
      error:
        example: false
        type: boolean
      errorresponse:
        $ref: '#/definitions/app.ErrorResponse'
      message:
        example: I am alive !!
        type: string
      seceretsavailable:
        example: "yes"
        type: string
    type: object
  app.PaymentMethodRequest:
    properties:
      CallerApp:
        example: mcp
        type: string
      FirstName:
        example: John
        type: string
      LastName:
        example: Doe
        type: string
      PaymentMethodType:
        example: ach
        type: string
      UserID:
        example: "125"
        type: string
      paymentMethodDetails:
        $ref: '#/definitions/app.Paymentmethoddetails'
    type: object
  app.PaymentMethodResponse:
    properties:
      errordetails:
        $ref: '#/definitions/app.ErrorResponse'
      paymentmethodid:
        example: 123456789
        type: integer
      status:
        example: ACCEPTED
        type: string
    type: object
  app.PaymentPreferenceDetail:
    properties:
      AutoPayPreference:
        example: true
        type: boolean
      PaymentMethodId:
        example: 1
        type: integer
      Split:
        example: 80
        type: integer
    type: object
  app.PaymentPreferenceRequest:
    properties:
      PaymentRequestType:
        example: enterpriserental
        type: string
      ProductIdentifier:
        example: ALL
        type: string
      ProductSubType:
        example: auto
        type: string
      ProductType:
        example: ppa
        type: string
      UserId:
        example: "125"
        type: string
      payIn:
        items:
          $ref: '#/definitions/app.PaymentPreferenceDetail'
        type: array
      payOut:
        items:
          $ref: '#/definitions/app.PaymentPreferenceDetail'
        type: array
    type: object
  app.PaymentPreferenceResponse:
    properties:
      AccountValidationDate:
        example: "2021-01-01T00:00:00Z"
        type: string
      AutoPayPreference:
        example: true
        type: boolean
      Last4AccountIdentifier:
        example: "6789"
        type: string
      PaymentExtendedData:
        example: '{}'
        type: string
      PaymentMethodStatus:
        example: ACTIVE
        type: string
      RoutingNumber:
        example: "023456789"
        type: string
      Split:
        example: 80
        type: integer
      Status:
        example: ACCEPTED
        type: string
      WalletAccess:
        example: true
        type: boolean
      WalletStatus:
        example: true
        type: boolean
      accountIdentifier:
        type: string
      errorDetails:
        $ref: '#/definitions/app.ErrorResponse'
      paymentMethodType:
        type: string
    type: object
  app.PaymentPreferenceResponseList:
    properties:
      PaymentPreferences:
        items:
          $ref: '#/definitions/app.PaymentPreferenceResponse'
        type: array
      errorDetails:
        $ref: '#/definitions/app.ErrorResponse'
      status:
        type: string
    type: object
  app.Paymentmethoddetails:
    properties:
      ACHAccountType:
        example: checking
        type: string
      AccountNumber:
        example: "123456789"
        type: string
      CVV:
        example: "123"
        type: string
      CardExpirationMonth:
        example: "12"
        type: string
      CardExpirationYear:
        example: "2022"
        type: string
      CardNumber:
        example: "123456789"
        type: string
      NickName:
        example: John's Account
        type: string
      RoutingNumber:
        example: "023456789"
        type: string
      ZipCode:
        example: "12345"
        type: string
    type: object
  app.StorePaymentPreferenceResponse:
    properties:
      errordetails:
        $ref: '#/definitions/app.ErrorResponse'
      paymentpreferenceid:
        example: 123456789
        type: integer
      status:
        example: ACCEPTED
        type: string
    type: object
info:
  contact:
    email: GPP@geico.com
    name: Payment Platform Dev Team
  description: Responsible for interacting with GEICO Payment Platform Vault.
  title: GEICO Payment Platform Vault API
  version: "1.0"
paths:
  /decrypt:
    get:
      description: This API decrypts the value passed in the query parameter.
      parameters:
      - description: Value to be decrypted
        in: query
        name: value
        required: true
        type: string
      produces:
      - application/json
      responses:
        default:
          description: Status of the request, errorDetails will be only displayed
            in case of 4XX or 5XX error
          schema:
            type: string
      summary: Get decrypted value.
      tags:
      - PaymentVault
  /encrypt:
    get:
      description: This API encrypts the value passed in the query parameter.
      parameters:
      - description: Value to be encrypted
        in: query
        name: value
        required: true
        type: string
      produces:
      - application/json
      responses:
        default:
          description: Status of the request, errorDetails will be only displayed
            in case of 4XX or 5XX error
          schema:
            type: string
      summary: Get encrypted value.
      tags:
      - PaymentVault
  /health/ready:
    get:
      description: This API returns the Health Check Status of the application
      produces:
      - application/json
      responses:
        default:
          description: Health Check Results
          schema:
            $ref: '#/definitions/app.HealthCheckResponse'
      summary: HealthCheck of the application
      tags:
      - HealthCheck
  /paymentmethod:
    post:
      consumes:
      - application/json
      description: This API adds a payment methods, with account related details,
        for a user, in GEICO Payment Platform.
      parameters:
      - description: Payment Preference method request body that needs to be added
        in: body
        name: PaymentMethodRequest
        required: true
        schema:
          $ref: '#/definitions/app.PaymentMethodRequest'
      produces:
      - application/json
      responses:
        default:
          description: Status of payment method request, errorDetails will be only
            displayed in case of 4XX or 5XX error
          schema:
            $ref: '#/definitions/app.PaymentMethodResponse'
      summary: Add Payment method.
      tags:
      - PaymentVault
  /paymentpreference:
    get:
      description: This API returns payment preference from GEICO Payment Platform,
        for a user, based on product, transactionType and Payment Request Type.
      parameters:
      - description: UserId from identity
        in: query
        name: userId
        required: true
        type: string
      - description: TransactionType (eg.PayIn,PayOut)
        in: query
        name: transactionType
        required: true
        type: string
      - description: PaymentRequestType(eg.CustomerChoice,insuranceautoauctions)
        in: query
        name: paymentRequestType
        required: true
        type: string
      - description: Product identifier eg.Policy Number
        in: query
        name: productIdentifier
        required: true
        type: string
      produces:
      - application/json
      responses:
        default:
          description: Status of get payment preference request, errorDetails will
            be only displayed in case of 4XX or 5XX error
          schema:
            $ref: '#/definitions/app.PaymentPreferenceResponseList'
      summary: Get Payment Preference.
      tags:
      - PaymentVault
    post:
      consumes:
      - application/json
      description: This API adds payment preference for payin and payout, for a user
        in GEICO Payment Platform.
      parameters:
      - description: Payment Preference Request body that needs to be added
        in: body
        name: PaymentPreferenceRequest
        required: true
        schema:
          $ref: '#/definitions/app.PaymentPreferenceRequest'
      produces:
      - application/json
      responses:
        default:
          description: Status of payment preference request, errorDetails will be
            only displayed in case of 4XX or 5XX error
          schema:
            $ref: '#/definitions/app.StorePaymentPreferenceResponse'
      summary: Add Payment Preference.
      tags:
      - PaymentVault
swagger: "2.0"